Adds show and index routes to
-Adds wor-paginate gem for pagination
-Adds active model serializers gem for responses serializatio
-Adds books controller with routes show and index
-Moves api_controller from app/controllers/ to app/controllers/api/v1/
-Adds 2 book serializer and a rent serializer to comply with api documentation
