+++ Gemfile
@@ -37,6 +37,9 @@ gem 'versionist'
 
 gem 'jwt'
 
+gem 'active_model_serializers', '~> 0.10.0'
+gem 'wor-paginate'
+
 group :development, :test do
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
   gem 'byebug', platforms: [:mri, :mingw, :x64_mingw]
