+++ app/controllers/api/v1/sessions_controller.rb
@@ -2,7 +2,7 @@
 
 module Api
   module V1
-    class SessionsController < ApiController
+    class SessionsController < Api::V1::ApiController
       skip_before_action :current_user, :authenticate_request, except: %i[renew invalidate_all]
 
       def create
