+++ app/serializers/book_show_serializer.rb
@@ -1,0 +1,8 @@
+class BookShowSerializer < ActiveModel::Serializer
+   attributes :id, :author, :title, :publisher, :year, :genre
+   attribute :image, key: :image_url
+
+   has_many :rent, key: :actual_rent, serializer: BookRentSerializer
+
+
+end
