+++ config/routes.rb
@@ -2,6 +2,8 @@ Rails.application.routes.draw do
   devise_for :users
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
   api_version(module: 'api/v1', path: { value: 'api/v1' }, defaults: { format: :json }) do
+      resources :books, only: [:index, :show]
+
       resources :users do
           collection do
               resources :sessions, only: [:create] do
